% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countries.R
\name{get_countrycode}
\alias{get_countrycode}
\title{Get country codes for a vector of country names}
\usage{
get_countrycode(name)
}
\arguments{
\item{name}{A character vector with country names}
}
\value{
A character vector with country codes
}
\description{
Returns the corresponding country code for a vector of country names.
Most countries are identified using alpha-3 (ISO 3166) codes.
Exceptions are Yugoslavia, 'Serbia and Montenegro', and 'Rest of the World',
which have custom codes.
}
\details{
Yugoslavia and 'Serbia and Montenegro' are countries that no longer exist.
'Rest of the World' is a special case with its own country code, ROW.

Uses \link[countrycode:codelist]{countrycode::codelist}.
}
\examples{
countries <- c("Sweden", "Italy", "Yugoslavia", "Rest of the World")

get_countrycode(countries)

}
\seealso{
\code{\link[=get_countryname]{get_countryname()}}
}
