% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countries.R
\name{get_countryname}
\alias{get_countryname}
\title{Get country names from country codes}
\usage{
get_countryname(code)
}
\arguments{
\item{code}{A character vector with country codes}
}
\value{
A character vector with country names
}
\description{
Returns the corresponding country name for a vector of country codes.
Most countries are identified using alpha-3 (ISO 3166) codes.
Exceptions are Yugoslavia, 'Serbia and Montenegro', and 'Rest of the World',
which have custom codes.
}
\details{
Yugoslavia and 'Serbia and Montenegro' are countries that no longer exist.
'Rest of the World' is a special case with its own country code, ROW.

Uses \link[countrycode:codelist]{countrycode::codelist}.
}
\examples{
country_codes <- c("SWE", "ITA", "YUG", "ROW")

get_countryname(country_codes)

}
\seealso{
\code{\link[=get_countrycode]{get_countrycode()}}
}
