% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/years.R
\name{get_years}
\alias{get_years}
\title{Get years pertaining to Eurovision Song Contest}
\usage{
get_years()
}
\value{
a list of named numeric vectors
}
\description{
The function returns a list with vectors of years pertaining to Eurovision
Song Contest. The list contains the following vectors:
}
\details{
\itemize{
\item \code{all}          - All years from contest start up until current year
\item \code{cancelled}    - Years when the event was cancelled
\item \code{one_round}    - Years that only had a Final round
\item \code{two_round}    - Years that had one Semi-Final and one Final round
\item \code{three_round}  - Years that had two Semi-Finals and one Final round
}
}
\examples{
# Get years that have had all three rounds
esc_years <- get_years()
esc_years$three_round

# Get years when the event was cancelled
esc_years$cancelled

}
