% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countries.R
\name{is_valid_country}
\alias{is_valid_country}
\title{Determine if a country name or code is valid within Eurovision}
\usage{
is_valid_country(x)
}
\arguments{
\item{x}{A character vector containing either country names or codes}
}
\value{
A logical vector indicating which inputs are valid
}
\description{
This function checks if each input is a valid country name or code.
}
\details{
Validity is determined by checking if the input can be converted to a
country code or name using the functions \code{\link[=get_countrycode]{get_countrycode()}} and
\code{\link[=get_countryname]{get_countryname()}}.
}
\examples{
is_valid_country("Sweden")

country_mix <- c("Sweden", "ITA", "Artist", NA, "ROW")

is_valid_country(country_mix)

}
