% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resc-package.R
\docType{package}
\name{resc-package}
\alias{resc}
\alias{resc-package}
\title{resc: Toolbox For Playing With Eurovision Data}
\description{
A collection of tools specifically for playing with Eurovision Song Contest (ESC) data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Phiido/resc}
  \item Report bugs at \url{https://github.com/Phiido/resc/issues}
}

}
\author{
\strong{Maintainer}: Thomas Roosdorp \email{thomas@roosdorp.se} (\href{https://orcid.org/0009-0002-5146-2777}{ORCID}) [copyright holder]

}
\keyword{internal}
