% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countries.R
\name{get_countrycode}
\alias{get_countrycode}
\title{Get country codes for a vector of country names}
\usage{
get_countrycode(name)
}
\arguments{
\item{name}{A character vector with country names}
}
\value{
A character vector with country codes
}
\description{
Finds country code for most countries using alpha-3 (ISO 3166) codes for most countries, and
custom code for the others.
}
\details{
Yugoslavia and 'Serbia and Montenegro' are exceptions to alpha-3 as they are countries
that no longer exists. In addition, 'Rest of the World' voting has its own
custom country code of ROW.

Uses \link[countrycode:codelist]{countrycode::codelist}.
}
\examples{
countries <- c('Sweden', 'Italy', 'Yugoslavia', 'Rest of the World')

get_countrycode(countries)

}
