% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countries.R
\name{get_countryname}
\alias{get_countryname}
\title{Get country names from country codes}
\usage{
get_countryname(code)
}
\arguments{
\item{code}{A character vector with country codes}
}
\value{
A character vector with country names
}
\description{
Checks for countries participated in Eurovision Song Contest. Finds country
name for most countries using alpha-3 (ISO 3166) codes, and custom code
for the others.
}
\details{
Yugoslavia and 'Serbia and Montenegro' are exceptions to alpha-3 as they are countries
that no longer exists. In addition, 'Rest of the World' voting has its own
country code of ROW.

Uses \link[countrycode:codelist]{countrycode::codelist}.
}
\examples{
country_codes <- c('SWE', 'ITA', 'YUG', 'ROW')

get_countryname(country_codes)

}
