% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/years.R
\name{get_years}
\alias{get_years}
\title{Get years pertaining to Eurovision Song Contest}
\usage{
get_years()
}
\value{
a list of numeric vectors
}
\description{
Builds a list with potential sets of years starting from contest start in
1956 up until the current year.
}
\details{
The following are possible to subset:
\itemize{
\item \code{all}          - Returns all years from contest start up until current year
\item \code{cancelled}    - Returns years when the event was cancelled
\item \code{one_round}    - Returns years that only had a Final round
\item \code{two_round}    - Returns years that had one Semi-Final and one Final round
\item \code{three_round}  - Returns years that had two Semi-Finals and one Final round
}

See examples for detailed usage.
}
\examples{
esc_years <- get_years()

# Which years have had all three rounds?
esc_years$three_round

# Which years were cancelled?
esc_years$cancelled

}
