% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countries.R
\name{is_valid_country}
\alias{is_valid_country}
\title{Is a country name or code valid within Eurovision}
\usage{
is_valid_country(x)
}
\arguments{
\item{x}{A character vector containing either country names or codes}
}
\value{
A logical vector
}
\description{
Is a country name or code valid within Eurovision
}
\examples{
is_valid_country('Sweden')

country_mix <- c('Sweden', 'ITA', 'Artist', NA, 'ROW')

is_valid_country(country_mix)

}
\seealso{
\code{\link[=get_countrycode]{get_countrycode()}} \code{\link[=get_countryname]{get_countryname()}}
}
